(hopefully) Fix testing error related to missed imports
